backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
langchain==0.0.350
openai==1.3.5
scikit-learn==1.4.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.7.4
httpx==0.24.1 